--- data/news.json
@@ -1,3 +1,0 @@
-{
-  
-}
