+++ index.html
@@ -3,13 +3,13 @@
 <head>
   <meta charset="UTF-8">
   <title>뉴스 스탠드</title>
-  <link rel="stylesheet" href="index.css">
+  <link rel="stylesheet" href="src/css/index.css">
 </head>
 <body>
   <div class="basic-display">
     <div class="top-container">
       <div class="top-container__title">
-        <img class="top-container__icon" src="img/newsStandLogo.svg" />
+        <img class="top-container__icon" src="src/img/newsStandLogo.svg" />
         <span class="top-container__title-text">뉴스스탠드</span>
       </div>
       <div class="top-container__date">
@@ -27,16 +27,16 @@
           <span class="press-container__subscribed-press">내가 구독한 언론사</span>
         </li>
         <li class="press-container__view-menu">
-          <img class="press-container__view-icon" src="img/list-view.svg" />
-          <img class="press-container__view-icon" src="img/grid-view.svg" />
+          <img class="press-container__view-icon" src="src/img/list-view.svg" />
+          <img class="press-container__view-icon" src="src/img/grid-view.svg" />
         </li>
       </ul>
       <div class="press-container__logo-table"></div>
-      <img class="press-container__left-arrow" src="img/LeftButton.svg" />
-      <img class="press-container__right-arrow" src="img/RightButton.svg" />
+      <img class="press-container__left-arrow" src="src/img/LeftButton.svg" />
+      <img class="press-container__right-arrow" src="src/img/RightButton.svg" />
     </div>
   </div>
-  <script src="runIndex.js"></script>
+  <script type="module" src="src/js/renderIndex.js"></script>
 </body>
 </html>
