+++ src/data/pressLogoTable.json
@@ -1,0 +1,1 @@
+{"width":932,"height":1554,"cellCountPerPage":24,"cells":[{"top":0,"left":0,"width":155.33333333333334,"height":97.125},{"top":0,"left":-155.33333333333334,"width":155.33333333333334,"height":97.125},{"top":0,"left":-310.6666666666667,"width":155.33333333333334,"height":97.125},{"top":0,"left":-466,"width":155.33333333333334,"height":97.125},{"top":0,"left":-621.3333333333334,"width":155.33333333333334,"height":97.125},{"top":0,"left":-776.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-97.125,"left":0,"width":155.33333333333334,"height":97.125},{"top":-97.125,"left":-155.33333333333334,"width":155.33333333333334,"height":97.125},{"top":-97.125,"left":-310.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-97.125,"left":-466,"width":155.33333333333334,"height":97.125},{"top":-97.125,"left":-621.3333333333334,"width":155.33333333333334,"height":97.125},{"top":-97.125,"left":-776.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-194.25,"left":0,"width":155.33333333333334,"height":97.125},{"top":-194.25,"left":-155.33333333333334,"width":155.33333333333334,"height":97.125},{"top":-194.25,"left":-310.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-194.25,"left":-466,"width":155.33333333333334,"height":97.125},{"top":-194.25,"left":-621.3333333333334,"width":155.33333333333334,"height":97.125},{"top":-194.25,"left":-776.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-291.375,"left":0,"width":155.33333333333334,"height":97.125},{"top":-291.375,"left":-155.33333333333334,"width":155.33333333333334,"height":97.125},{"top":-291.375,"left":-310.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-291.375,"left":-466,"width":155.33333333333334,"height":97.125},{"top":-291.375,"left":-621.3333333333334,"width":155.33333333333334,"height":97.125},{"top":-291.375,"left":-776.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-388.5,"left":0,"width":155.33333333333334,"height":97.125},{"top":-388.5,"left":-155.33333333333334,"width":155.33333333333334,"height":97.125},{"top":-388.5,"left":-310.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-388.5,"left":-466,"width":155.33333333333334,"height":97.125},{"top":-388.5,"left":-621.3333333333334,"width":155.33333333333334,"height":97.125},{"top":-388.5,"left":-776.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-485.625,"left":0,"width":155.33333333333334,"height":97.125},{"top":-485.625,"left":-155.33333333333334,"width":155.33333333333334,"height":97.125},{"top":-485.625,"left":-310.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-485.625,"left":-466,"width":155.33333333333334,"height":97.125},{"top":-485.625,"left":-621.3333333333334,"width":155.33333333333334,"height":97.125},{"top":-485.625,"left":-776.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-582.75,"left":0,"width":155.33333333333334,"height":97.125},{"top":-582.75,"left":-155.33333333333334,"width":155.33333333333334,"height":97.125},{"top":-582.75,"left":-310.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-582.75,"left":-466,"width":155.33333333333334,"height":97.125},{"top":-582.75,"left":-621.3333333333334,"width":155.33333333333334,"height":97.125},{"top":-582.75,"left":-776.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-679.875,"left":0,"width":155.33333333333334,"height":97.125},{"top":-679.875,"left":-155.33333333333334,"width":155.33333333333334,"height":97.125},{"top":-679.875,"left":-310.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-679.875,"left":-466,"width":155.33333333333334,"height":97.125},{"top":-679.875,"left":-621.3333333333334,"width":155.33333333333334,"height":97.125},{"top":-679.875,"left":-776.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-777,"left":0,"width":155.33333333333334,"height":97.125},{"top":-777,"left":-155.33333333333334,"width":155.33333333333334,"height":97.125},{"top":-777,"left":-310.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-777,"left":-466,"width":155.33333333333334,"height":97.125},{"top":-777,"left":-621.3333333333334,"width":155.33333333333334,"height":97.125},{"top":-777,"left":-776.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-874.125,"left":0,"width":155.33333333333334,"height":97.125},{"top":-874.125,"left":-155.33333333333334,"width":155.33333333333334,"height":97.125},{"top":-874.125,"left":-310.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-874.125,"left":-466,"width":155.33333333333334,"height":97.125},{"top":-874.125,"left":-621.3333333333334,"width":155.33333333333334,"height":97.125},{"top":-874.125,"left":-776.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-971.25,"left":0,"width":155.33333333333334,"height":97.125},{"top":-971.25,"left":-155.33333333333334,"width":155.33333333333334,"height":97.125},{"top":-971.25,"left":-310.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-971.25,"left":-466,"width":155.33333333333334,"height":97.125},{"top":-971.25,"left":-621.3333333333334,"width":155.33333333333334,"height":97.125},{"top":-971.25,"left":-776.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-1068.375,"left":0,"width":155.33333333333334,"height":97.125},{"top":-1068.375,"left":-155.33333333333334,"width":155.33333333333334,"height":97.125},{"top":-1068.375,"left":-310.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-1068.375,"left":-466,"width":155.33333333333334,"height":97.125},{"top":-1068.375,"left":-621.3333333333334,"width":155.33333333333334,"height":97.125},{"top":-1068.375,"left":-776.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-1165.5,"left":0,"width":155.33333333333334,"height":97.125},{"top":-1165.5,"left":-155.33333333333334,"width":155.33333333333334,"height":97.125},{"top":-1165.5,"left":-310.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-1165.5,"left":-466,"width":155.33333333333334,"height":97.125},{"top":-1165.5,"left":-621.3333333333334,"width":155.33333333333334,"height":97.125},{"top":-1165.5,"left":-776.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-1262.625,"left":0,"width":155.33333333333334,"height":97.125},{"top":-1262.625,"left":-155.33333333333334,"width":155.33333333333334,"height":97.125},{"top":-1262.625,"left":-310.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-1262.625,"left":-466,"width":155.33333333333334,"height":97.125},{"top":-1262.625,"left":-621.3333333333334,"width":155.33333333333334,"height":97.125},{"top":-1262.625,"left":-776.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-1359.75,"left":0,"width":155.33333333333334,"height":97.125},{"top":-1359.75,"left":-155.33333333333334,"width":155.33333333333334,"height":97.125},{"top":-1359.75,"left":-310.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-1359.75,"left":-466,"width":155.33333333333334,"height":97.125},{"top":-1359.75,"left":-621.3333333333334,"width":155.33333333333334,"height":97.125},{"top":-1359.75,"left":-776.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-1456.875,"left":0,"width":155.33333333333334,"height":97.125},{"top":-1456.875,"left":-155.33333333333334,"width":155.33333333333334,"height":97.125},{"top":-1456.875,"left":-310.6666666666667,"width":155.33333333333334,"height":97.125},{"top":-1456.875,"left":-466,"width":155.33333333333334,"height":97.125},{"top":-1456.875,"left":-621.3333333333334,"width":155.33333333333334,"height":97.125},{"top":-1456.875,"left":-776.6666666666667,"width":155.33333333333334,"height":97.125}]}
