+++ src/js/initialize.js
@@ -1,0 +1,9 @@
+import initializeNews from "./initializeNews.js";
+import initializePressTable from "./initializePressTable.js"
+
+const initialize = () => {
+  initializePressTable();
+  initializeNews();
+}
+
+initialize();
