+++ src/js/initializeNews.js
@@ -2,11 +2,6 @@ import fs from "fs";
 import axios from "axios";
 import cheerio from "cheerio";
 
-const PRESS_LOGOS_WIDTH = 932;
-const PRESS_LOGOS_HEIGHT = 1554;
-const ROW_COUNT = 16;
-const COLUMN_COUNT = 6;
-const CELL_COUNT_PER_PAGE = 24;
 const NO_ELEMENT = 0;
 const INCREMENT = 1;
 const URL =
@@ -22,47 +17,6 @@ const parseStringFromDate = (date) => {
   return `${date.getFullYear()}-${
     date.getMonth() + INCREMENT
   }-${date.getDate()}`;
-};
-
-const addWidthHeightSettings = (json) => {
-  json.width = PRESS_LOGOS_WIDTH;
-  json.height = PRESS_LOGOS_HEIGHT;
-  json.cellCountPerPage = CELL_COUNT_PER_PAGE;
-};
-
-const addCellsIntoTable = (json) => {
-  const cells = [];
-  const pressLogoWidth = PRESS_LOGOS_WIDTH / COLUMN_COUNT;
-  const pressLogoHeight = PRESS_LOGOS_HEIGHT / ROW_COUNT;
-  const rows = Array.from({ length: ROW_COUNT });
-  const columns = Array.from({ length: COLUMN_COUNT });
-
-  rows.forEach((_, rowIndex) => {
-    columns.forEach((_, columnIndex) => {
-      cells.push({
-        top: -rowIndex * pressLogoHeight,
-        left: -columnIndex * pressLogoWidth,
-        width: pressLogoWidth,
-        height: pressLogoHeight,
-      });
-    });
-  });
-
-  json.cells = cells;
-};
-
-const initializePressTable = () => {
-  const path = "data/pressLogoTable.json";
-  const pressLogoTableFile = fs.readFileSync(path, "utf-8");
-  const pressLogoTable =
-    pressLogoTableFile === "" ? {} : JSON.parse(pressLogoTableFile);
-
-  if (isEmpty(pressLogoTable)) {
-    addWidthHeightSettings(pressLogoTable);
-    addCellsIntoTable(pressLogoTable);
-  }
-
-  fs.writeFileSync(path, JSON.stringify(pressLogoTable));
 };
 
 const crawlNewsTitles = async (json) => {
@@ -83,7 +37,7 @@ const crawlNewsTitles = async (json) => {
 };
 
 const initializeNews = async () => {
-  const path = "data/news.json";
+  const path = "src/data/news.json";
   const newsFile = fs.readFileSync(path, "utf-8");
   const newsTitles = newsFile === "" ? {} : JSON.parse(newsFile);
   const currentDate = new Date();
@@ -99,5 +53,4 @@ const initializeNews = async () => {
   fs.writeFileSync(path, JSON.stringify(newsTitles));
 };
 
-initializePressTable();
-initializeNews();
+export default initializeNews;
