+++ src/js/renderCurrentDate.js
@@ -1,0 +1,28 @@
+const WEEK = [
+  "일요일",
+  "월요일",
+  "화요일",
+  "수요일",
+  "목요일",
+  "금요일",
+  "토요일",
+];
+const INCREMENT = 1;
+const CHAR_COUNT = 2;
+
+const renderCurrentDate = () => {
+  const currentDateTag = document.querySelector(".top-container__date-text");
+  const currentDate = new Date();
+  const year = currentDate.getFullYear();
+  const month = currentDate.getMonth() + INCREMENT;
+  const date = currentDate.getDate();
+  const day = WEEK[currentDate.getDay()];
+
+  currentDateTag.innerHTML = `${year}. ${month
+    .toString()
+    .padStart(CHAR_COUNT, "0")}. ${date
+    .toString()
+    .padStart(CHAR_COUNT, "0")}. ${day}`;
+};
+
+export default renderCurrentDate;
